Move state implementation to step
