+++ src/auto_scheduler/transform_step.cc
@@ -29,6 +29,7 @@
 #include <tvm/te/operation.h>
 
 #include <utility>
+#include <vector>
 
 #include "loop_state.h"
 #include "utils.h"
@@ -62,6 +63,16 @@ ReorderStep::ReorderStep(int stage_id, const Array<Integer>& after_ids) {
   data_ = std::move(node);
 }
 
+void ReorderStepNode::ApplyToState(State* state) const {
+  const Stage& stage = (*state)->stages[stage_id];
+  Array<Iterator> iters;
+  for (auto x : after_ids) {
+    iters.push_back(stage->iters[x]);
+  }
+  state->CopyOnWrite()->stages.Set(
+      stage_id, Stage(stage->op, stage->op_type, iters, stage->compute_at, stage->attrs));
+}
+
 void ReorderStepNode::ApplyToSchedule(Array<te::Stage>* stages,
                                       StageToAxesMap* stage_to_axes) const {
   auto stage = (*stages)[stage_id];
@@ -79,7 +90,7 @@ void ReorderStepNode::ApplyToSchedule(Array<te::Stage>* stages,
   stages->Set(stage_id, std::move(stage));
 }
 
-String ReorderStepNode::PrintAsPythonAPI(Array<te::Stage>* stages,
+String ReorderStepNode::ApplyToPythonAPI(Array<te::Stage>* stages,
                                          StageToAxesMap* stage_to_axes) const {
   const auto& stage = (*stages)[stage_id];
   std::stringstream ss;
@@ -106,6 +117,23 @@ ComputeAtStep::ComputeAtStep(int stage_id, int target_stage_id, int target_iter_
   data_ = std::move(node);
 }
 
+void ComputeAtStepNode::ApplyToState(State* state) const {
+  const Stage& stage = (*state)->stages[stage_id];
+
+  // Remove the bound information of each iterator since they may not be accurate after
+  // compute at
+  Array<Iterator> new_iters;
+  for (const Iterator& it : stage->iters) {
+    new_iters.push_back(Iterator(it->name, Range(), it->iter_kind, it->annotation));
+  }
+
+  StateNode* pstate = state->CopyOnWrite();
+  pstate->stages.Set(stage_id, Stage(stage->op, stage->op_type, std::move(new_iters),
+                                     ComputeAtKind::kIter, stage->attrs));
+  // Update attach map
+  pstate->attach_map.SetComputeAtIter(stage_id, target_stage_id, target_iter_id);
+}
+
 void ComputeAtStepNode::ApplyToSchedule(Array<te::Stage>* stages,
                                         StageToAxesMap* stage_to_axes) const {
   te::Stage stage = (*stages)[stage_id];
@@ -116,7 +144,7 @@ void ComputeAtStepNode::ApplyToSchedule(Array<te::Stage>* stages,
   stages->Set(stage_id, std::move(stage));
 }
 
-String ComputeAtStepNode::PrintAsPythonAPI(Array<te::Stage>* stages,
+String ComputeAtStepNode::ApplyToPythonAPI(Array<te::Stage>* stages,
                                            StageToAxesMap* stage_to_axes) const {
   std::stringstream ss;
   const auto& stage = (*stages)[stage_id];
@@ -134,6 +162,23 @@ ComputeRootStep::ComputeRootStep(int stage_id) {
   data_ = std::move(node);
 }
 
+void ComputeRootStepNode::ApplyToState(State* state) const {
+  const Stage& stage = (*state)->stages[stage_id];
+
+  // Remove the bound information of each iterator since they may not be accurate after
+  // compute root
+  Array<Iterator> new_iters;
+  for (const Iterator& it : stage->iters) {
+    new_iters.push_back(Iterator(it->name, Range(), it->iter_kind, it->annotation));
+  }
+
+  StateNode* pstate = state->CopyOnWrite();
+  pstate->stages.Set(stage_id, Stage(stage->op, stage->op_type, std::move(new_iters),
+                                     ComputeAtKind::kRoot, stage->attrs));
+  // Update attach map
+  pstate->attach_map.DeleteStage(stage_id);
+}
+
 void ComputeRootStepNode::ApplyToSchedule(Array<te::Stage>* stages,
                                           StageToAxesMap* stage_to_axes) const {
   auto stage = (*stages)[stage_id];
@@ -141,7 +186,7 @@ void ComputeRootStepNode::ApplyToSchedule(Array<te::Stage>* stages,
   stages->Set(stage_id, std::move(stage));
 }
 
-String ComputeRootStepNode::PrintAsPythonAPI(Array<te::Stage>* stages,
+String ComputeRootStepNode::ApplyToPythonAPI(Array<te::Stage>* stages,
                                              StageToAxesMap* stage_to_axes) const {
   std::stringstream ss;
   const auto& stage = (*stages)[stage_id];
@@ -157,6 +202,24 @@ ComputeInlineStep::ComputeInlineStep(int stage_id) {
   data_ = std::move(node);
 }
 
+void ComputeInlineStepNode::ApplyToState(State* state) const {
+  const Stage& stage = (*state)->stages[stage_id];
+
+  // Check the validity of compute_inline
+  for (size_t i = 0; i < stage->iters.size(); ++i) {
+    CHECK_EQ((*state)->attach_map->iter_to_attached_stages.count(std::make_pair(stage_id, i)), 0)
+        << "Invalid compute_inline: There are some other stages that are attached to the "
+        << "target stage";
+  }
+
+  StateNode* pstate = state->CopyOnWrite();
+  auto new_stage = pstate->stages[stage_id];
+  new_stage.CopyOnWrite()->compute_at = ComputeAtKind::kInlined;
+  pstate->stages.Set(stage_id, std::move(new_stage));
+  // Update attach map
+  pstate->attach_map.DeleteStage(stage_id);
+}
+
 void ComputeInlineStepNode::ApplyToSchedule(Array<te::Stage>* stages,
                                             StageToAxesMap* stage_to_axes) const {
   auto stage = (*stages)[stage_id];
@@ -164,7 +227,7 @@ void ComputeInlineStepNode::ApplyToSchedule(Array<te::Stage>* stages,
   stages->Set(stage_id, std::move(stage));
 }
 
-String ComputeInlineStepNode::PrintAsPythonAPI(Array<te::Stage>* stages,
+String ComputeInlineStepNode::ApplyToPythonAPI(Array<te::Stage>* stages,
                                                StageToAxesMap* stage_to_axes) const {
   std::stringstream ss;
   const auto& stage = (*stages)[stage_id];
@@ -174,6 +237,86 @@ String ComputeInlineStepNode::PrintAsPythonAPI(Array<te::Stage>* stages,
 }
 
 /********** Split **********/
+// common part for SplitStep, FollowSplitStep, and FollowFusedSplitStep
+Array<Iterator> ApplySplitToState(State* state, int stage_id, int iter_id,
+                                  const Array<Optional<Integer>>& lengths, bool inner_to_outer) {
+  const Stage& stage = (*state)->stages[stage_id];
+  const Iterator& it = stage->iters[iter_id];
+  size_t old_iter_size = stage->iters.size();
+  bool concrete = true;
+
+  Optional<PrimExpr> tosplit_min, tosplit_extent;
+  if (it->range.defined()) {
+    tosplit_min = it->range->min;
+    tosplit_extent = it->range->extent;
+  } else {
+    tosplit_min = NullOpt;
+    tosplit_extent = NullOpt;
+  }
+
+  Array<Iterator> outs;
+  for (size_t i = 0; i < lengths.size(); ++i) {
+    Optional<Integer> l;
+    String name;
+    if (inner_to_outer) {
+      l = lengths[lengths.size() - i - 1];
+      name = it->name + "." + std::to_string(lengths.size() - i);
+    } else {
+      l = lengths[i];
+      name = it->name + "." + std::to_string(i);
+    }
+    Iterator res;
+    if (l && tosplit_min && tosplit_extent) {
+      res = Iterator(name, Range::FromMinExtent(tosplit_min.value(), l.value()), it->iter_kind,
+                     IteratorAnnotation::kNone);
+      tosplit_min = Integer(0);
+      tosplit_extent = indexdiv(tosplit_extent.value() + l.value() - 1, l.value());
+    } else {
+      res = Iterator(name, Range(), it->iter_kind, IteratorAnnotation::kNone);
+      tosplit_min = NullOpt;
+      tosplit_extent = NullOpt;
+      concrete = false;
+    }
+    outs.push_back(std::move(res));
+  }
+
+  Range range;
+  if (tosplit_min && tosplit_extent) {
+    range = Range::FromMinExtent(tosplit_min.value(), tosplit_extent.value());
+  }
+  if (inner_to_outer) {
+    outs.push_back(Iterator(it->name + ".0", range, it->iter_kind, IteratorAnnotation::kNone));
+    // Reverse the Iterator array
+    Array<Iterator> temp(outs.rbegin(), outs.rend());
+    outs = std::move(temp);
+  } else {
+    outs.push_back(Iterator(it->name + "." + std::to_string(lengths.size()), range, it->iter_kind,
+                            IteratorAnnotation::kNone));
+  }
+
+  Array<Iterator> new_iters;
+  new_iters.insert(new_iters.end(), stage->iters.begin(), stage->iters.begin() + iter_id);
+  new_iters.insert(new_iters.end(), outs.begin(), outs.end());
+  new_iters.insert(new_iters.end(), stage->iters.begin() + iter_id + 1, stage->iters.end());
+
+  StateNode* pstate = state->CopyOnWrite();
+  pstate->stages.Set(stage_id,
+                     Stage(stage->op, stage->op_type, new_iters, stage->compute_at, stage->attrs));
+  pstate->concrete &= concrete;
+
+  // Two vectors are used to represent the iterator relation before and after split
+  // The original iterators in AttachMap will be updated with the new iterators
+  std::vector<IterKey> from_iters;
+  std::vector<IterKey> to_iters;
+  for (size_t i = iter_id; i < old_iter_size; ++i) {
+    from_iters.emplace_back(stage_id, i);
+    to_iters.emplace_back(stage_id, i + lengths.size());
+  }
+  pstate->attach_map.UpdateIters(from_iters, to_iters);
+
+  return outs;
+}
+
 Array<IterVar> ApplySplitToSchedule(Array<te::Stage>* stages, StageToAxesMap* stage_to_axes,
                                     int stage_id, int iter_id,
                                     const Array<Optional<Integer>>& lengths, bool inner_to_outer) {
@@ -262,12 +405,16 @@ SplitStep::SplitStep(int stage_id, int iter_id, Optional<PrimExpr> extent,
   data_ = std::move(node);
 }
 
+Array<Iterator> SplitStepNode::ApplyToState(State* state) const {
+  return ApplySplitToState(state, stage_id, iter_id, lengths, inner_to_outer);
+}
+
 Array<IterVar> SplitStepNode::ApplyToSchedule(Array<te::Stage>* stages,
                                               StageToAxesMap* stage_to_axes) const {
   return ApplySplitToSchedule(stages, stage_to_axes, stage_id, iter_id, lengths, inner_to_outer);
 }
 
-String SplitStepNode::PrintAsPythonAPI(Array<te::Stage>* stages,
+String SplitStepNode::ApplyToPythonAPI(Array<te::Stage>* stages,
                                        StageToAxesMap* stage_to_axes) const {
   return PrintSplitAsPythonAPI(stages, stage_to_axes, stage_id, iter_id, lengths, inner_to_outer);
 }
@@ -281,6 +428,85 @@ FuseStep::FuseStep(int stage_id, const Array<Integer>& fused_ids) {
   }
   node->fused_ids = fused_ids;
   data_ = std::move(node);
+}
+
+Iterator FuseStepNode::ApplyToState(State* state) const {
+  const Stage& stage = (*state)->stages[stage_id];
+  size_t old_iter_size = static_cast<int>(stage->iters.size());
+
+  String new_name;
+  PrimExpr new_extent = 1;
+  IteratorKind new_iter_kind = IteratorKind::kSpecial;
+
+  for (size_t i = 0; i < fused_ids.size(); ++i) {
+    if (i > 0) {
+      CHECK_EQ(fused_ids[i]->value, fused_ids[i - 1]->value + 1);
+    }
+
+    if (i != fused_ids.size() - 1) {
+      const auto& iter_to_attached_stage = (*state)->attach_map->iter_to_attached_stages;
+      if (iter_to_attached_stage.find(std::make_pair(stage_id, fused_ids[i])) !=
+          iter_to_attached_stage.end()) {
+        LOG(FATAL) << "Invalid Fuse. Trying to fuse iterators that have been attached by some "
+                   << "stages. State before fusion:\n"
+                   << (*state);
+      }
+    }
+
+    const Iterator& it = stage->iters[fused_ids[i]];
+    new_name = new_name + it->name + "@";
+
+    if (it->range.defined() && new_extent.defined()) {
+      new_extent = new_extent * it->range->extent;
+    } else {
+      new_extent = PrimExpr();
+    }
+
+    if (i == 0) {
+      new_iter_kind = it->iter_kind;
+    } else {
+      if (new_iter_kind != it->iter_kind) {
+        new_iter_kind = IteratorKind::kMixed;
+      }
+    }
+  }
+
+  Range range;
+  if (new_extent.defined()) {
+    range = Range::FromMinExtent(0, new_extent);
+  }
+  Iterator new_it = Iterator(new_name, range, new_iter_kind, IteratorAnnotation::kNone);
+  Array<Iterator> new_iters;
+  new_iters.insert(new_iters.end(), stage->iters.begin(), stage->iters.begin() + fused_ids.front());
+  new_iters.push_back(new_it);
+  new_iters.insert(new_iters.end(), stage->iters.begin() + fused_ids.back() + 1,
+                   stage->iters.end());
+
+  StateNode* pstate = state->CopyOnWrite();
+  pstate->stages.Set(stage_id,
+                     Stage(stage->op, stage->op_type, new_iters, stage->compute_at, stage->attrs));
+
+  // Two vectors are used to represent the iterator relation before and after fuse
+  // The original iterators in AttachMap will be updated with the new iterators
+  std::vector<IterKey> from_iters;
+  std::vector<IterKey> to_iters;
+  const size_t begin_id = fused_ids.front(), end_id = fused_ids.back();
+  for (size_t i = 0; i < old_iter_size; ++i) {
+    if (i <= begin_id) {
+      continue;
+    } else if (i > end_id) {
+      // move forward
+      from_iters.emplace_back(stage_id, i);
+      to_iters.emplace_back(stage_id, i - end_id + begin_id);
+    } else {
+      // move to the fused id
+      from_iters.emplace_back(stage_id, i);
+      to_iters.emplace_back(stage_id, begin_id);
+    }
+  }
+  pstate->attach_map.UpdateIters(from_iters, to_iters);
+
+  return new_it;
 }
 
 IterVar FuseStepNode::ApplyToSchedule(Array<te::Stage>* stages,
@@ -305,7 +531,7 @@ IterVar FuseStepNode::ApplyToSchedule(Array<te::Stage>* stages,
   return fused_axis;
 }
 
-String FuseStepNode::PrintAsPythonAPI(Array<te::Stage>* stages,
+String FuseStepNode::ApplyToPythonAPI(Array<te::Stage>* stages,
                                       StageToAxesMap* stage_to_axes) const {
   const auto& stage = (*stages)[stage_id];
   std::stringstream to_fuse;
@@ -333,6 +559,18 @@ AnnotationStep::AnnotationStep(int stage_id, int iter_id, IteratorAnnotation ann
   node->iter_id = iter_id;
   node->annotation = ann;
   data_ = std::move(node);
+}
+
+Iterator AnnotationStepNode::ApplyToState(State* state) const {
+  const Stage& stage = (*state)->stages[stage_id];
+  Iterator it = stage->iters[iter_id];
+
+  CHECK(it->annotation == IteratorAnnotation::kNone);
+  Iterator new_it = Iterator(it->name, it->range, it->iter_kind, annotation);
+  Stage new_stage = stage;
+  new_stage.CopyOnWrite()->iters.Set(iter_id, new_it);
+  state->CopyOnWrite()->stages.Set(stage_id, std::move(new_stage));
+  return new_it;
 }
 
 void AnnotationStepNode::ApplyToSchedule(Array<te::Stage>* stages,
@@ -370,7 +608,7 @@ void AnnotationStepNode::ApplyToSchedule(Array<te::Stage>* stages,
   stages->Set(stage_id, std::move(stage));
 }
 
-String AnnotationStepNode::PrintAsPythonAPI(Array<te::Stage>* stages,
+String AnnotationStepNode::ApplyToPythonAPI(Array<te::Stage>* stages,
                                             StageToAxesMap* stage_to_axes) const {
   std::stringstream ss;
   const auto& stage = (*stages)[stage_id];
